P4

+++ src/main/java/repository/DemandeRepository.java
@@ -2,8 +2,7 @@ package repository;
 
 import model.Demande;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-public interface DemandeRepository extends JpaRepository<Demande, Integer> {
-    
+public interface DemandeRepository extends JpaRepository<Demande, Long> {
+
 }
